Added base setup for testing
